bollmaker: fix short position order
